+++ app/models/cmd/user.go
@@ -14,10 +14,6 @@ type UnblockUser struct {
 	UserID int
 }
 
-type UntrustUser struct {
-	UserID int
-}
-
 type RegenerateAPIKey struct {
 	Result string
 }