+++ public/services/actions/post.ts
@@ -21,7 +21,6 @@ export interface SearchPostsParams {
   noTags?: boolean
   myPosts?: boolean
   statuses?: string[]
-  moderation?: string
 }
 
 export interface PostPagination {
@@ -42,7 +41,6 @@ const postSearchURL = (params: SearchPostsParams, page?: number): string => {
     query: params.query,
     view: params.view ? normalizePostView(params.view) : undefined,
     limit: params.limit,
-    moderation: params.moderation,
     notags: params.noTags ? "true" : undefined,
     myposts: params.myPosts ? "true" : undefined,
   })
@@ -127,36 +125,4 @@ export const createPost = async (title: string, description: string, attachments
 
 export const updatePost = async (postNumber: number, title: string, description: string, attachments: ImageUpload[]): Promise<Result> => {
   return http.put(`/api/v1/posts/${postNumber}`, { title, description, attachments }).then(http.event("post", "update"))
-}
-
-export const approvePost = async (postID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/posts/${postID}/approve`).then(http.event("post", "approve"))
-}
-
-export const declinePost = async (postID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/posts/${postID}/decline`).then(http.event("post", "decline"))
-}
-
-export const approveComment = async (commentID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/comments/${commentID}/approve`).then(http.event("comment", "approve"))
-}
-
-export const declineComment = async (commentID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/comments/${commentID}/decline`).then(http.event("comment", "decline"))
-}
-
-export const approvePostAndVerify = async (postID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/posts/${postID}/approve-and-verify`).then(http.event("post", "approve-and-verify"))
-}
-
-export const declinePostAndBlock = async (postID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/posts/${postID}/decline-and-block`).then(http.event("post", "decline-and-block"))
-}
-
-export const approveCommentAndVerify = async (commentID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/comments/${commentID}/approve-and-verify`).then(http.event("comment", "approve-and-verify"))
-}
-
-export const declineCommentAndBlock = async (commentID: number): Promise<Result> => {
-  return http.post(`/api/v1/admin/moderation/comments/${commentID}/decline-and-block`).then(http.event("comment", "decline-and-block"))
 }